Refactor UI tests
Move friend classes to Helpers
Register UI tests in the separate header
